# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is based on the Philippine Standard Geographic Code (PSGC), which is a systematic classification and coding of geographic areas in the Philippines. Its units of classification are based on the four well-established levels of geographical-political subdivisions of the country, including the region, the province, the municipality/city, and the barangay.",
        "title": "Philippine Standard Geographic Code (PSGC) API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/barangays": {
            "get": {
                "description": "get Barangays",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barangays"
                ],
                "summary": "Show list of Barangays",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedBarangay"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/barangays/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barangays"
                ],
                "summary": "Show a Barangay",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barangay psgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Barangay"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citi_muni": {
            "get": {
                "description": "get Cities/Municipalities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities/Municipalities"
                ],
                "summary": "Show list of Cities/Municipalities",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedCityMuni"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/citi_muni/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities/Municipalities"
                ],
                "summary": "Show a City/Municipality",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City/Municipality PsgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CityMuni"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cities": {
            "get": {
                "description": "get Cities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Show list of Cities",
                "parameters": [
                    {
                        "type": "string",
                        "example": "filter",
                        "description": "This is filter for all the field in the  object",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "example": 1000,
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedCityMuni"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cities/{psgc_code}": {
            "get": {
                "description": "get string by PsgcCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cities"
                ],
                "summary": "Show a City",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City PsgcCode",
                        "name": "psgc_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CityMuni"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Item Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Barangay": {
            "type": "object",
            "properties": {
                "city_muni_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "psgc_code": {
                    "type": "string"
                }
            }
        },
        "CityMuni": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prov_code": {
                    "type": "string"
                },
                "psgc_code": {
                    "type": "string"
                }
            }
        },
        "MetaData": {
            "type": "object",
            "properties": {
                "item_count": {
                    "type": "integer",
                    "example": 1000
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 1000
                },
                "total_items": {
                    "type": "integer",
                    "example": 10000
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "PaginatedBarangay": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Barangay"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/MetaData"
                }
            }
        },
        "PaginatedCityMuni": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CityMuni"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/MetaData"
                }
            }
        },
...
    },
...
}